refactor : selenium 명령어 클래스 분리
